+++ cw-orch-daemon/src/senders/cosmos.rs
@@ -132,6 +132,28 @@ impl Wallet {
         self.options.clone()
     }
 
+    /// Replaces the private key that the [CosmosSender] is using with key derived from the provided 24-word mnemonic.
+    /// If you want more control over the derived private key, use [Self::set_private_key]
+    pub fn set_mnemonic(&mut self, mnemonic: impl Into<String>) -> Result<(), DaemonError> {
+        let secp = Secp256k1::new();
+
+        let pk = PrivateKey::from_words(
+            &secp,
+            &mnemonic.into(),
+            0,
+            self.options.hd_index.unwrap_or(0),
+            self.chain_info.network_info.coin_type,
+        )?;
+        self.set_private_key(pk);
+        Ok(())
+    }
+
+    /// Replaces the private key the sender is using
+    /// You can use a mnemonic to overwrite the key using [Self::set_mnemonic]
+    pub fn set_private_key(&mut self, private_key: PrivateKey) {
+        self.private_key = private_key
+    }
+
     pub fn set_authz_granter(&mut self, granter: &Addr) {
         self.options.authz_granter = Some(granter.to_owned());
     }